Ta i bruk bidrag beregning

+++ src/main/kotlin/no/nav/bidrag/bidragskalkulator/controller/dto/BeregningResultatDto.kt
@@ -1,5 +1,7 @@
 package no.nav.bidrag.bidragskalkulator.controller.dto
 
+import java.math.BigDecimal
+
 data class BeregningResultatDto(
-    val resultat: Double
+    val resultat: BigDecimal
 )
